requirements.txt
matplotlib==3.3.4
numpy==1.19.5
pandas==1.1.5
scikit-learn==0.24.1
scipy==1.5.4
tensorboard==2.5.0
#tensorflow==2.5.0
tf-keras-vis==0.5.5
tqdm==4.58.0
sentencepiece==0.1.95
pydot==1.4.2
tensorflow-text==2.5.0
